+++ stem-explorer-ng/package.json
@@ -12,7 +12,6 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^3.0.0-beta.0",
     "@angular/animations": "~10.0.9",
     "@angular/cdk": "~8.2.3",
     "@angular/common": "~10.0.9",
@@ -21,7 +20,6 @@
     "@angular/fire": "^6.0.2",
     "@angular/flex-layout": "^10.0.0-beta.32",
     "@angular/forms": "~10.0.9",
-    "@angular/google-maps": "^9.2.4",
     "@angular/localize": "^10.0.9",
     "@angular/material": "~8.2.3",
     "@angular/platform-browser": "~10.0.9",
