+++ stem-explorer-ng/src/app/shared/shared.module.ts
@@ -3,8 +3,6 @@ import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { ReactiveFormsModule } from '@angular/forms';
-import { GoogleMapsModule } from '@angular/google-maps';
-import { ZXingScannerModule } from '@zxing/ngx-scanner';
 import { GeolocationService } from '../../locations/services/geolocation.service';
 import { ButtonComponent } from './components/button/button.component';
 import { CardComponent } from './components/card/card.component';
@@ -15,22 +13,18 @@ import { FormFieldComponent } from './components/form-field/form-field.component
 import { InputComponent } from './components/input/input.component';
 import { ProfilePictureComponent } from './components/profile-picture/profile-picture.component';
 import { MaterialModule } from './material.module';
-import { FilterLocationsPipe } from './pipes/filter-locations.pipe';
 import { FilterPipe } from './pipes/filter.pipe';
 import { LargeDistancePipe } from './pipes/large-distance.pipe';
 import { SortByPipe } from './pipes/sort-by.pipe';
 import { TruncatePipe } from './pipes/truncate.pipe';
 import { ApiService } from './services/api.service';
-import { ImageService } from './services/image.service';
 import { StemColorsService } from './services/stem-colors.service';
 import { ImageService } from './services/image.service';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { ReactiveFormsModule } from '@angular/forms';
-import { ProfileReminderService } from './services/profile-reminder.service';
 import { FilterLocationsPipe } from './pipes/filter-locations.pipe';
-import { VisibleDirective } from './directives/visible.directive';
-import { MessageService } from './services/message.service';
 import { ConfirmDialogComponent } from './components/confirm-dialog/confirm-dialog.component';
+import { VisibleDirective } from './directives/visible.directive';
+import { ProfileReminderService } from './services/profile-reminder.service';
+import { MessageService } from './services/message.service';
 
 @NgModule({
   declarations: [
@@ -78,7 +72,6 @@ import { ConfirmDialogComponent } from './components/confirm-dialog/confirm-dial
     InputComponent,
     SortByPipe,
     MaterialModule,
-    GoogleMapsModule,
     ProfilePictureComponent,
     FlexLayoutModule,
     ReactiveFormsModule,
