+++ stem-explorer-ng/src/locations/components/map/map.component.ts
@@ -1,255 +1,267 @@
-// import { Component, OnInit, ViewChild, ChangeDetectorRef, ChangeDetectionStrategy, ElementRef, AfterViewInit } from '@angular/core';
-// import { MapInfoWindow } from '@angular/google-maps';
-// import { MapMarker } from '@angular/google-maps/map-marker/map-marker';
-// import { MatDialog } from '@angular/material/dialog';
-// import { Store } from '@ngxs/store';
-// import { GoogleTagManagerService } from 'angular-google-tag-manager';
-// import { map } from 'rxjs/operators';
-// import { ChallengeDialogType } from 'src/app/shared/enums/challenge-dialog-type.enum';
-// import { StemColours } from 'src/app/shared/enums/stem-colours.enum';
-// import { Location, LocationChallenge } from 'src/locations/models/location';
-// import { GeolocationService } from 'src/locations/services/geolocation.service';
-// import { MapConfigService } from 'src/locations/services/map-config.service';
-// import { LoadLocationsData } from 'src/locations/store/locations.actions';
-// import { LocationsState } from 'src/locations/store/locations.state';
-// import { ChallengeDialogComponent } from '../challenge-dialog/challenge-dialog.component';
-// import { LargeCategoryIcons } from 'src/app/shared/enums/large-category-icons.enum';
-// import { Router } from '@angular/router';
-// import { LocationApiService } from 'src/locations/services/locations-api.service';
-// import { Filter } from 'src/locations/models/filter';
+import {
+  OnInit,
+  AfterViewInit,
+  ViewChild,
+  ElementRef,
+  Component,
+  OnDestroy,
+  ComponentFactoryResolver,
+  ApplicationRef,
+  Injector,
+  ViewContainerRef,
+  TemplateRef,
+} from '@angular/core';
+import { MapConfigService } from 'src/locations/services/map-config.service';
+import { Location, LocationChallenge } from '../../models/location';
+import { Filter } from 'src/locations/models/filter';
+import { FilterLocationsPipe } from 'src/app/shared/pipes/filter-locations.pipe';
+import { GeolocationService } from 'src/locations/services/geolocation.service';
+import { map } from 'rxjs/operators';
+import { GoogleTagManagerService } from 'angular-google-tag-manager';
+import { LargeCategoryIcons } from 'src/app/shared/enums/large-category-icons.enum';
+import { StemColours } from 'src/app/shared/enums/stem-colours.enum';
+import { MatDialog } from '@angular/material';
+import { ChallengeDialogComponent } from '../challenge-dialog/challenge-dialog.component';
+import { MapIcon } from 'src/locations/models/map-icons.constant';
+import { Store } from '@ngxs/store';
+import { LocationsState } from 'src/locations/store/locations.state';
+import { LoadLocationsData } from 'src/locations/store/locations.actions';
+import { DomPortalOutlet, TemplatePortal } from '@angular/cdk/portal';
+import { VisitedHomepage } from 'src/app/store/last-homepage/last-homepage.actions';
 
-// // @Component({
-// //   selector: 'app-map',
-// //   templateUrl: './map.component.html',
-// //   styleUrls: ['./map.component.scss'],
-// // })
-// export class MapComponent implements OnInit, AfterViewInit {
-//   public filter: Filter;
-//   @ViewChild(MapInfoWindow, { static: false }) infoWindow: MapInfoWindow;
+@Component({
+  selector: 'app-map',
+  templateUrl: './map.component.html',
+  styleUrls: ['./map.component.scss'],
+  providers: [FilterLocationsPipe]
+})
+export class MapComponent implements OnInit, AfterViewInit, OnDestroy {
+  @ViewChild('mapContainer', { static: false }) gmap: ElementRef;
+  @ViewChild('infoWindow', { static: false }) infoWindow: TemplateRef<unknown>;
+  map: google.maps.Map;
+  markers = new Map<Location, google.maps.Marker>();
 
+  filter: Filter;
+  locations: Location[];
+  location: Location;
+  distance: string;
+  userLocation: google.maps.LatLngLiteral;
+  userLocationLat: number;
+  userLocationLng: number;
+  Colour = StemColours;
+  Icon = LargeCategoryIcons;
+  locationAccess = false;
+  locationsSubscription: any;
 
-//   locations: Location[] = [];
-//   zoom = 16;
-//   center: google.maps.LatLngLiteral;
-//   userLocation: google.maps.LatLngLiteral;
-//   options: google.maps.MapOptions;
-//   location: Location;
-//   Colour = StemColours;
-//   Icon = LargeCategoryIcons;
-//   locationAccess = false;
-//   tilesLoaded = false;
-//   distance: string;
+  infoW: google.maps.InfoWindow;
+  userMarker: google.maps.Marker;
+  portal: TemplatePortal<any>;
 
-//   constructor(
-//     private store: Store,
-//     private dialog: MatDialog,
-//     private gtmService: GoogleTagManagerService,
-//     private mapConfig: MapConfigService,
-//     private geolocation: GeolocationService,
-//     private api: LocationApiService,
-//     private router: Router
-//   ) {
-//     this.options = this.mapConfig.mapOptions();
-//     this.geolocation.getMapCentre().then(pos => {
-//       this.center = {
-//         lat: pos.lat,
-//         lng: pos.lng
-//       };
-//     });
-//     this.geolocation.getPosition().then(pos => {
-//       if (pos) {
-//         this.userLocation = {
-//           lat: pos.lat,
-//           lng: pos.lng
-//         };
-//       }
-//     });
+  constructor(
+    private mapConfig: MapConfigService,
+    private store: Store,
+    private filterLocations: FilterLocationsPipe,
+    private geolocation: GeolocationService,
+    private gtmService: GoogleTagManagerService,
+    private dialog: MatDialog,
+    private componentFactoryResolver: ComponentFactoryResolver,
+    private appRef: ApplicationRef,
+    private defaultInjector: Injector,
+    private viewContainerRef: ViewContainerRef,
+  ) {
+    this.geolocation.getPosition().then(pos => {
+      if (pos) {
+        this.userLocationLat = pos.lat;
+        this.userLocationLng = pos.lng;
+        this.userLocation = {
+          lat: pos.lat,
+          lng: pos.lng
+        };
 
-//     this.locationAccess = !navigator.geolocation;
-//   }
+        if (this.userMarker) {
+          this.userMarker.setPosition(pos);
+        }
+      }
+    });
 
-  
-//   @ViewChild('mapContainer', { static: false }) gmap: ElementRef;
-//   map: google.maps.Map;
-//   lat = 40.73061;
-//   lng = -73.935242;
+    this.locationAccess = !navigator.geolocation;
+  }
 
-//   // markers = [
-//   //   {
-//   //     position: new google.maps.LatLng(40.73061, 73.935242),
-//   //     map: this.map,
-//   //     title: "Marker 1"
-//   //   },
-//   //   {
-//   //     position: new google.maps.LatLng(32.06485, 34.763226),
-//   //     map: this.map,
-//   //     title: "Marker 2"
-//   //   }
-//   // ];
+  ngOnInit(): void {
+    this.store.dispatch(new VisitedHomepage());
+    this.getLocations();
+  }
 
-//   coordinates = new google.maps.LatLng(-37.6854709, 176.1673285);
+  ngAfterViewInit(): void {
+    this.mapInit();
+  }
 
-//   // marker = new google.maps.Marker({
-//   //   position: this.coordinates,
-//   //   map: this.map,
-//   //   title: "Hello World!"
-//   // });
+  ngOnDestroy(): void {
+    this.locationsSubscription?.unsubscribe();
+  }
 
-//   mapOptions: google.maps.MapOptions = {
-//     center: this.coordinates,
-//     zoom: 16
-//   };
+  trackLocations(_: number, item: Location) {
+    return item?.uid;
+  }
 
-//   ngAfterViewInit(): void {
-//     this.mapInitializer();
-//   }
+  trackChallenges(_: number, item: LocationChallenge) {
+    return item?.challengeId;
+  }
 
-//   mapInitializer(): void {
-//     // this.map = new google.maps.Map(this.gmap.nativeElement, this.mapOptions);
+  mapInit(): void {
+    this.map = new google.maps.Map(this.gmap.nativeElement, this.mapConfig.mapOptions());
+    this.setMapMarkers();
+  }
 
-//     // this.marker.addListener("click", () => {
-//     //   const infoWindow = new google.maps.InfoWindow({
-//     //     content: this.marker.getTitle()
-//     //   });
-//     //   infoWindow.open(this.marker.getMap(), this.marker);
-//     // });
+  filterChanged(filter: Filter) {
+    this.filter = filter;
+    this.setMapMarkers();
+  }
 
-//     // this.marker.setMap(this.map);
+  clickOnMarker(marker, location: Location): void {
+    console.warn('click on marker')
+    this.getDistanceToLocation(location.position);
+    this.location = location;
 
-//     // this.loadAllMarkers();
-//   }
+    this.infoW?.close();
+    this.portal?.detach();
 
-//   loadAllMarkers(): void {
-//     // this.markers.forEach(markerInfo => {
-//     //   const marker = new google.maps.Marker({
-//     //     ...markerInfo
-//     //   });
+    this.infoW = new google.maps.InfoWindow({
+      content: '<div id="info-window-container"></div>'
+    });
+    this.infoW.open(marker.getMap(), marker);
 
-//     //   const infoWindow = new google.maps.InfoWindow({
-//     //     content: marker.getTitle()
-//     //   });
+    this.infoW.addListener('domready', () => {
+      const el = this.gmap.nativeElement as HTMLElement;
+      const container = el.querySelector('#info-window-container');
+      // Dom Portals allow us to display Angular components inside
+      // non-Angular DOM elements
+      const portalOutlet = new DomPortalOutlet(
+        container,
+        this.componentFactoryResolver,
+        this.appRef,
+        this.defaultInjector
+      );
+      const portal = new TemplatePortal(this.infoWindow, this.viewContainerRef);
+      portal.attach(portalOutlet);
+      this.portal = portal;
+    });
 
-//     //   marker.addListener("click", () => {
-//     //     infoWindow.open(marker.getMap(), marker);
-//     //   });
+    this.addGtmTag('open location info', location.name);
+  }
 
-//     //   marker.setMap(this.map);
-//     // });
-//   }
+  /**
+   * Method fired when user clicked on a challenge button in the Location info window. Triggers the Challenge dialog box.
+   * @param location location data object
+   * @param challenge individual challenge
+   * @todo have a contact/links object in the location info, so we can send through that and only location name. Currently
+   * if location has multiple challenges they also get sent through.
+   */
+  openChallenge(location: Location, challenge: LocationChallenge): void {
+    this.dialog.open(ChallengeDialogComponent, {
+      data: {
+        challenge,
+        location,
+      },
+      panelClass: 'app-dialog',
+    });
+    this.addGtmTag('open challenge info', challenge.challengeTitle);
+  }
 
-//   ngOnInit() {
-//     this.getLocations();
-//   }
+  /**
+   * Gets all locations from store
+   */
+  private getLocations(): void {
+    this.store.dispatch(new LoadLocationsData());
 
-//   trackLocations(_: number, item: Location) {
-//     return item?.uid;
-//   }
+    this.locationsSubscription = this.store
+      .select(LocationsState.locations)
+      .subscribe((res) => {
+        this.locations = res;
+        this.setMapMarkers();
+      });
+  }
 
-//   trackChallenges(_: number, item: LocationChallenge) {
-//     return item?.challengeId;
-//   }
+  private setMapMarkers(): void {
+    if (!this.locations || !this.filter) { return; }
 
-//   filterChanged(filter: Filter) {
-//     this.filter = filter;
-//   }
+    const filtered = this.filterLocations.transform(this.locations, this.filter);
 
-//   /**
-//    * Method for when a user clicks on a map marker on the map. Shows the location details and the list of challenges
-//    * and gets the distance from user's current position to the location.
-//    * @param marker google MapMarker object
-//    * @param location location of mapmarker
-//    */
-//   click(marker: MapMarker, location: Location): void {
-//     this.getDistanceToLocation(location.position);
-//     this.location = location;
-//     this.infoWindow.open(marker);
-//     this.addGtmTag('open location info', location.name);
-//   }
+    // Delete markers that are no longer shown
+    this.markers.forEach((marker, loc) => {
+      const stillVisible = filtered.indexOf(loc) >= 0;
+      if (!stillVisible) {
+        marker.setMap(null);
+        this.markers.delete(loc);
+      }
+    });
 
-//   /**
-//    * Method fired when user clicked on a challenge button in the Location info window. Triggers the Challenge dialog box.
-//    * @param location location data object
-//    * @param challenge individual challenge
-//    * @todo have a contact/links object in the location info, so we can send through that and only location name. Currently
-//    * if location has multiple challenges they also get sent through.
-//    */
-//   openChallenge(location: Location, challenge: LocationChallenge): void {
-//     this.dialog.open(ChallengeDialogComponent, {
-//       data: {
-//         challenge,
-//         location,
-//         dialogType: ChallengeDialogType.Preview,
-//       },
-//       panelClass: 'app-dialog',
-//     });
-//     this.addGtmTag('open challenge info', challenge.challengeTitle);
-//   }
+    // Add new markers
+    filtered.forEach(loc => {
+      if (this.markers.has(loc)) {
+        // Don't create duplicate markers
+        return;
+      }
 
-//   /**
-//    * Method for getting the map marker icon. Checks if the location has more then one challenge, and if so will show a different icon
-//    * but otherwise will show an icon for that challenge category.
-//    * @param location location data object
-//    */
-//   getMarkerOptions(location: Location): google.maps.MarkerOptions {
-//     let iconUrl = this.mapConfig.mapMarkerIcons(4);
-//     if (location.challengeCount === 1) {
-//       iconUrl = this.mapConfig.mapMarkerIcons(location.locationChallenges[0].challengeCategory);
-//     }
-//     return {
-//       icon: {
-//         url: `/assets/icons/${iconUrl}`,
-//         scaledSize: new google.maps.Size(30, 48)
-//       },
-//       animation: google.maps.Animation.DROP
-//     };
-//   }
+      const marker = new google.maps.Marker({
+        position: new google.maps.LatLng(loc.position),
+        title: loc.name,
+        map: this.map,
+        icon: {
+          url: this.returnMapIcon(loc.locationChallenges),
+          scaledSize: new google.maps.Size(30, 48)
+        }
+      });
 
-//   /**
-//    * Gets an icon to show the users position on the map.
-//    */
-//   userMarkerPoint(): google.maps.MarkerOptions {
-//     return {
-//       icon: {
-//         url: '/assets/icons/personMarker.png'
-//       }
-//     };
-//   }
+      marker.addListener('click', () => {
+        this.clickOnMarker(marker, loc);
+      });
 
-//   /**
-//    * Gets all locations from API
-//    */
-//   private getLocations(): void {
-//     this.api.getLocations().subscribe((res) => {
-//       this.locations = res;
-//     })
-//     // this.store.select(LocationsState.locations).pipe(map(res => {
-//     //   this.locations = res;
-//     // })).subscribe();
-//   }
+      this.markers.set(loc, marker);
+    });
 
-//   /**
-//    * Gets the distance to location. Resets distance to an empty string to prevent previous distance
-//    * from showing in the Info Window, and only assigns the distance text to it if current has a
-//    * current position. If user does not have a current position, then obs returns undefined and distance
-//    * remains as an empty string.
-//    * @param position the lat and lng object of a location
-//    */
-//   private getDistanceToLocation(position: google.maps.LatLngLiteral): void {
-//     this.distance = '';
+    if (!this.userMarker) {
+      this.userMarker = new google.maps.Marker({
+        position: new google.maps.LatLng(this.userLocationLat, this.userLocationLng),
+        map: this.map,
+        icon: '/assets/icons/personMarker.png'
+      });
+    }
+    this.userMarker.setMap(this.map);
 
-//     if (this.userLocation) {
-//       this.geolocation.getDistance(position, this.userLocation).pipe(
-//         map(res => this.distance = res)
-//       ).subscribe();
-//     }
-//   }
+    this.markers.forEach(m => m.setMap(this.map));
+  }
 
-//   /**
-//    * add tag to GTM on the card click
-//    * @param title challenge title
-//    */
-//   private addGtmTag(event: string, title: string): void {
-//     const gtmTag = { event, title };
-//     this.gtmService.pushTag(gtmTag);
-//   }
-// }
+  returnMapIcon(challenges: LocationChallenge[]) {
+    if (challenges.length > 1) {
+      return MapIcon[4];
+    }
+    return MapIcon[challenges[0].challengeCategory];
+  }
+
+  /**
+   * Gets the distance to location. Resets distance to an empty string to prevent previous distance
+   * from showing in the Info Window, and only assigns the distance text to it if current has a
+   * current position. If user does not have a current position, then obs returns undefined and distance
+   * remains as an empty string.
+   * @param position the lat and lng object of a location
+   */
+  private getDistanceToLocation(position: google.maps.LatLngLiteral): void {
+    this.distance = '';
+
+    if (this.userLocation) {
+      this.geolocation.getDistance(position, this.userLocation).pipe(
+        map(res => this.distance = res)
+      ).subscribe();
+    }
+  }
+
+  /**
+   * add tag to GTM on the card click
+   * @param title challenge title
+   */
+  private addGtmTag(event: string, title: string): void {
+    const gtmTag = { event, title };
+    this.gtmService.pushTag(gtmTag);
+  }
+}
