+++ stem-explorer-ng/tsconfig.app.json
@@ -1,7 +1,8 @@
 {
   "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "outDir": "./out-tsc/app"
+    "outDir": "./out-tsc/app",
+    "types": ["googlemaps"]
   },
   "files": [
     "src/main.ts",
